# server/src/docs/swagger.json
{
	"definitions": {
...
		"UserI": {
			"description": "",
			"properties": {
				"id": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"name": {
					"type": "string",
					"description": ""
				},
				"lastName": {
					"type": "string",
					"description": ""
				},
				"email": {
					"type": "string",
					"description": ""
				},
				"password": {
					"type": "string",
					"description": ""
				},
				"idDocumentType": {
					"type": "string",
					"description": ""
				},
				"idDocumentNumber": {
					"type": "integer",
					"format": "int32",
					"description": ""
				},
				"rol": {
					"type": "string",
					"description": ""
				},
				"status": {
					"type": "integer",
					"format": "int32",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"name",
				"lastName",
				"email",
				"idDocumentType",
				"idDocumentNumber",
				"rol",
				"status"
			]
		},
...
		"UserL": {
			"description": "",
			"properties": {
				"email": {
					"type": "string",
					"description": ""
				},
				"password": {
					"type": "string",
					"description": ""
				}
			},
			"type": "object",
			"required": [
				"email",
				"password"
			]
		}
	},
	"info": {
		"description": "Shoppy API.",
		"license": {
			"name": "ISC"
		},
		"title": "shoppy-api",
		"version": "1.0.0"
	},
	"paths": {
...
		"/api/users/q": {
			"get": {
				"operationId": "UserGetByEmail",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Retrieve an User.",
						"schema": {
							"$ref": "#/definitions/UserI"
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "User not found."
					}
				},
				"description": "Produce an user with a given email",
				"tags": [
					"Users"
				],
				"parameters": [
					{
						"description": "",
						"in": "query",
						"name": "email",
						"required": true,
						"type": "string"
					}
				]
			}
		},
		"/api/users/update": {
			"put": {
				"operationId": "UserUpdate",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Updates an User.",
						"schema": {
							"$ref": "#/definitions/UserI"
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "User not found."
					}
				},
				"description": "UPDATES an user",
				"tags": [
					"Users"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"entity": {
									"$ref": "#/definitions/UserI"
								},
								"decodedToken": {
									"$ref": "#/definitions/UserI"
								}
							}
						}
					}
				]
			}
		},
		"/api/users/signup": {
			"post": {
				"operationId": "UserPost",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Creates an User.",
						"schema": {
							"$ref": "#/definitions/UserI"
						}
					},
					"204": {
						"description": "No content"
					},
					"500": {
						"description": "Server Error."
					}
				},
				"description": "CREATES an user",
				"tags": [
					"Users"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"name": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserI"
						}
					}
				]
			}
		},
		"/api/users/login": {
			"post": {
				"operationId": "UserLogin",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Logins an user.",
						"schema": {
							"$ref": "#/definitions/UserI"
						}
					},
					"204": {
						"description": "No content"
					},
					"401": {
						"description": "User not found."
					}
				},
				"description": "Logins an user with a given email and password",
				"tags": [
					"Users"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "- {email: string, password: string}",
						"in": "body",
						"name": "user",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserL"
						}
					}
				]
			}
		},
		"/api/users/": {
			"delete": {
				"operationId": "UserDelete",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "User deleted.",
						"schema": {
							"$ref": "#/definitions/UserI"
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "User not found."
					}
				},
				"description": "DELETE an user",
				"tags": [
					"Users"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"entity": {
									"$ref": "#/definitions/UserI"
								},
								"decodedToken": {
									"$ref": "#/definitions/UserI"
								}
							}
						}
					}
				]
			},
			"post": {
				"operationId": "UserGetAll",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Retrieve a list of all Users.",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/UserI"
							}
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Users not found."
					}
				},
				"description": "Produce a list of all users",
				"tags": [
					"Users"
				],
				"consumes": [
					"application/json"
				],
				"parameters": [
					{
						"description": "",
						"in": "body",
						"required": true,
						"schema": {
							"type": "object",
							"properties": {
								"decodedToken": {
									"$ref": "#/definitions/UserI"
								}
							}
						}
					}
				]
			}
		},
		"/api/users/{id}": {
			"get": {
				"operationId": "UserGetById",
				"produces": [
					"application/json",
					"text/html"
				],
				"responses": {
					"200": {
						"description": "Retrieve an User.",
						"schema": {
							"$ref": "#/definitions/UserI"
						}
					},
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "User not found."
					}
				},
				"description": "GET an user with a given id",
				"tags": [
					"Users"
				],
				"parameters": [
					{
						"description": "",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "number",
						"format": "double"
					}
				]
			}
		}
	},
	"swagger": "2.0",
	"securityDefinitions": {
		"client secret": {
			"type": "apiKey",
			"name": "x-auth",
			"in": "header"
		}
	}
}